Added early stopping and queue management in update edge method and some fixes for optimization
The min_distance array keeps track of the shortest distances found during the update propagation. If a new distance calculated isn't an improvement over min_distance, the propagation for that path stops.
Nodes are only pushed back into the queue if the new distance is an actual improvement, reducing unnecessary computations